backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
# pip install openai-whisper --no-build-isolation
openai-whisper
python-dotenv
httpx
pydantic
numpy
sqlalchemy
google-auth
requests
email-validator
scipy
faster-whisper
# LLM providers
google-genai
# Optional: Ollama fallback (uncomment if needed)
# langchain
# langchain-ollama
# Optional: HuggingFace TTS fallback (uncomment if needed)
# transformers
# torch
